model/pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "KinjiKawaguchi", email = "kawakin0310@icloud.com" }
]
dependencies = [
    "pygithub>=2.3.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/model"]
